refactor: split monolithic commands.ts into individual command modules
The 1756-line commands.ts file was the largest file in the codebase.
This splits it into 15 focused modules under src/handlers/commands/:

- shared.ts: common utilities (getCommandLines, buildSessionOverviewLines, etc.)
- state.ts: Claude state/backlog parsing types and functions
- new.ts, status.ts, model.ts, switch.ts, resume.ts: core command handlers
- usage.ts: subscription usage tracking and formatting
- context.ts: context usage reporting
- looplogs.ts: run-loop log tail
- subturtle.ts: SubTurtle management + parsing utilities
- debug.ts: internal state inspection
- restart.ts: bot restart handler
- cron.ts: scheduled jobs listing
- index.ts: barrel re-export preserving all existing import paths

No consumer files needed updating — the barrel index.ts ensures
"./commands" imports resolve identically to the old single file.
Test file paths updated to reference commands/index.ts.
